Close #2591 clean up scalable milestone
