+++ app/controllers/spree/api/v2/storefront/accommodations/variants_controller.rb
@@ -2,15 +2,15 @@ module Spree
   module Api
     module V2
       module Storefront
-        class AccommodationsController < ::Spree::Api::V2::ResourceController
+        class Accommodations::VariantsController < ::Spree::Api::V2::ResourceController
           private
 
           # override
           def collection
-            @collection ||= SpreeCmCommissioner::Accommodations::Find.new(
+            @collection ||= SpreeCmCommissioner::Accommodations::FindVariant.new(
               from_date: params[:from_date]&.to_date,
               to_date: params[:to_date]&.to_date,
-              state_id: params[:state_id],
+              vendor_id: params[:accommodation_id],
               number_of_adults: params[:number_of_adults].to_i,
               number_of_kids: params[:number_of_kids].to_i
             ).execute
@@ -22,23 +22,18 @@ module Spree
           end
 
           # override
-          def allowed_sort_attributes
-            super << :min_price << :max_price
-          end
-
-          # override
           def resource_serializer
-            Spree::V2::Storefront::AccommodationSerializer
+            Spree::V2::Storefront::VariantSerializer
           end
 
           # override
           def collection_serializer
-            Spree::V2::Storefront::AccommodationSerializer
+            Spree::V2::Storefront::VariantSerializer
           end
 
           # override
           def required_schema
-            SpreeCmCommissioner::AccommodationRequestSchema
+            SpreeCmCommissioner::VariantRequestSchema
           end
         end
       end
