+++ lib/spree_cm_commissioner/version.rb
@@ -1,5 +1,5 @@
 module SpreeCmCommissioner
-  VERSION = '1.9.0'.freeze
+  VERSION = '1.10.0-pre'.freeze
 
   module_function
