+++ spec/request/spree/api/v2/storefront/accommodations/variants_request_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe Spree::Api::V2::Storefront::Novel::AccommodationsController, type: :request do
+RSpec.describe Spree::Api::V2::Storefront::Accommodations::VariantsController, type: :request do
   let(:state) { create(:state, id: 1) }
   let(:vendor) do
     create(:cm_vendor,
@@ -26,12 +26,12 @@ RSpec.describe Spree::Api::V2::Storefront::Novel::AccommodationsController, type
     end
   end
 
-  describe 'GET /api/v2/storefront/novel/accommodations' do
+  describe 'GET /api/v2/storefront/accommodations/:accommodation_id/variants' do
     let(:valid_params) do
       {
         from_date: from_date.to_s,
         to_date: to_date.to_s,
-        state_id: state.id,
+        accommodation_id: vendor.id,
         number_of_adults: 2,
         number_of_kids: 1
       }
@@ -39,28 +39,28 @@ RSpec.describe Spree::Api::V2::Storefront::Novel::AccommodationsController, type
 
     context 'with valid parameters' do
       before do
-        get '/api/v2/storefront/novel/accommodations', params: valid_params
+        get "/api/v2/storefront/accommodations/#{vendor.id}/variants", params: valid_params
       end
 
       it 'returns a successful response' do
         expect(response).to have_http_status(:ok)
       end
 
-      it 'returns a JSON response with vendor data' do
+      it 'returns a JSON response with variant data' do
         expect(json_response_body['data']).to be_present
         expect(json_response_body['data']).to be_an(Array)
-        expect(json_response_body['data'].first['type']).to eq('vendor')
-        expect(json_response_body['data'].first['id']).to eq(vendor.id.to_s)
+        expect(json_response_body['data'].first['type']).to eq('variant')
+        expect(json_response_body['data'].first['id']).to eq(variant.id.to_s)
         expect(json_response_body['data'].first['attributes']).to include(
-          'min_price' => vendor.min_price.to_s,
-          'max_price' => vendor.max_price.to_s
+          'sku' => variant.sku,
+          'price' => variant.price.to_s
         )
       end
     end
 
     context 'with missing required parameters' do
       before do
-        get '/api/v2/storefront/novel/accommodations', params: { state_id: state.id }
+        get "/api/v2/storefront/accommodations/#{vendor.id}/variants", params: { accommodation_id: vendor.id }
       end
 
       it 'returns a bad request response' do
@@ -69,42 +69,18 @@ RSpec.describe Spree::Api::V2::Storefront::Novel::AccommodationsController, type
       end
     end
 
-    context 'with sorting by min_price' do
-      let(:vendor2) do
+    context 'with no matching variants' do
+      let(:other_vendor) do
         create(:cm_vendor,
           default_state: state,
           primary_product_type: :accommodation,
-          state: :active,
-          min_price: 50,
-          max_price: 150
+          state: :active
         )
       end
-      let(:product) { create(:product, vendor: vendor, product_type: 'accommodation') }
-      let(:variant2) { create(:cm_variant, number_of_adults: 2, number_of_kids: 2, vendor: vendor2, product: product) }
+      let(:params_with_different_vendor) { valid_params.merge(accommodation_id: other_vendor.id) }
 
       before do
-        (from_date..to_date.prev_day).each do |date|
-          create(:cm_inventory_item,
-            variant: variant2,
-            inventory_date: date,
-            quantity_available: 5
-          )
-        end
-        get '/api/v2/storefront/novel/accommodations', params: valid_params.merge(sort: 'min_price')
-      end
-
-      it 'returns vendors sorted by min_price asc' do
-        expect(response).to have_http_status(:ok)
-        expect(json_response_body['data'].map { |d| d['id'] }).to eq([vendor.id.to_s, vendor2.id.to_s])
-      end
-    end
-
-    context 'with no matching vendors' do
-      let(:other_state) { create(:state, id: 2) }
-      let(:params_with_different_state) { valid_params.merge(state_id: other_state.id) }
-
-      before do
-        get '/api/v2/storefront/novel/accommodations', params: params_with_different_state
+        get "/api/v2/storefront/accommodations/#{other_vendor.id}/variants", params: params_with_different_vendor
       end
 
       it 'returns an empty collection' do
@@ -114,12 +90,11 @@ RSpec.describe Spree::Api::V2::Storefront::Novel::AccommodationsController, type
     end
   end
 
-  describe 'GET /api/v2/storefront/novel/accommodations/:id' do
+  describe 'GET /api/v2/storefront/accommodations/:accommodation_id/variants/:id' do
     let(:valid_params) do
       {
         from_date: from_date.to_s,
         to_date: to_date.to_s,
-        state_id: state.id,
         number_of_adults: 2,
         number_of_kids: 1
       }
@@ -127,35 +102,44 @@ RSpec.describe Spree::Api::V2::Storefront::Novel::AccommodationsController, type
 
     context 'with a valid ID' do
       before do
-        get "/api/v2/storefront/novel/accommodations/#{vendor.id}", params: valid_params
+        get "/api/v2/storefront/accommodations/#{vendor.id}/variants/#{variant.id}", params: valid_params
       end
 
       it 'returns a successful response' do
         expect(response).to have_http_status(:ok)
       end
 
-      it 'returns the specific vendor as JSON' do
+      it 'returns the specific variant as JSON' do
         expect(json_response_body['data']).to be_present
-        expect(json_response_body['data']['id']).to eq(vendor.id.to_s)
-        expect(json_response_body['data']['type']).to eq('vendor')
+        expect(json_response_body['data']['id']).to eq(variant.id.to_s)
+        expect(json_response_body['data']['type']).to eq('variant')
         expect(json_response_body['data']['attributes']).to include(
-          'min_price' => vendor.min_price.to_s,
-          'max_price' => vendor.max_price.to_s
+          'sku' => variant.sku,
+          'price' => variant.price.to_s
         )
       end
     end
 
     context 'with an invalid ID' do
-      let(:invalid_vendor) do
-        create(:vendor,
+      let(:other_vendor) do
+        create(:cm_vendor,
           default_state: create(:state, id: 2),
           primary_product_type: :accommodation,
           state: :active
         )
       end
+      let(:product) { create(:product, vendor: other_vendor, product_type: 'accommodation') }
+      let(:invalid_variant) { create(:cm_variant, number_of_adults: 2, number_of_kids: 2, vendor: other_vendor, product: product) }
 
       before do
-        get "/api/v2/storefront/novel/accommodations/#{invalid_vendor.id}", params: valid_params
+        (from_date..to_date.prev_day).each do |date|
+          create(:cm_inventory_item,
+            variant: invalid_variant,
+            inventory_date: date,
+            quantity_available: 5
+          )
+        end
+        get "/api/v2/storefront/accommodations/#{vendor.id}/variants/#{invalid_variant.id}", params: valid_params
       end
 
       it 'returns a not found response' do
